Exception handling - delete

+++ src/main/java/com/marcuscardosoteste/teste/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.marcuscardosoteste.teste.entities.User;
 import com.marcuscardosoteste.teste.repositories.UserRepository;
+import com.marcuscardosoteste.teste.services.exceptions.DatabaseException;
 import com.marcuscardosoteste.teste.services.exceptions.ResourceNotFoundException;
 @Service
 public class UserService {
@@ -29,8 +32,16 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+	    if (!repository.existsById(id)) {
+	        throw new ResourceNotFoundException("Usuario com id: " + id + " nao encontrado");
+	    }
+	    try {
+	        repository.deleteById(id);
+	    } catch (DataIntegrityViolationException e) {
+	        throw new DatabaseException(e.getMessage());
+	    }
 	}
+
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
